+++ src/collective/normalize_buildout/__init__.py
@@ -1,6 +1,1 @@
 # -*- coding: utf-8 -*-
-"""Init and utils."""
-
-from zope.i18nmessageid import MessageFactory
-
-_ = MessageFactory('collective.normalize_buildout')