+++ src/collective/normalize_buildout/tests/test_script.py
@@ -42,3 +42,42 @@ a=1
 '''
 
         self.assertEqual(expected, output.read())
+
+    def test_regression1(self):
+        cfg = self.given_a_file_in_test_dir('buildout.cfg', '\n'.join([
+            '[sources]',
+            '# xxx',
+            '# yyy',
+            'a = git http...',
+        ]))
+        output = StringIO()
+
+        sort(file(cfg), output)
+        output.seek(0)
+
+        expected = '''[sources]
+# xxx
+# yyy
+a = git http...
+'''
+
+        self.assertEqual(expected, output.read())
+
+    def test_regression2(self):
+        cfg = self.given_a_file_in_test_dir('buildout.cfg', '''
+[filter]
+extra-field-types =
+            <charFilter class="solr.PatternReplaceCharFilterFactory" pattern="(/)+$" replacement=""/>
+'''.strip())  # NOQA
+        output = StringIO()
+
+        sort(file(cfg), output)
+        output.seek(0)
+
+        expected = '''
+[filter]
+extra-field-types =
+            <charFilter class="solr.PatternReplaceCharFilterFactory" pattern="(/)+$" replacement=""/>
+'''.strip()  # NOQA
+
+        self.assertEqual(expected, output.read())