+++ CHANGES.rst
@@ -5,7 +5,11 @@ Changelog
 0.1.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Remove unneeded dependency.
+  [do3cc]
+
+- Fix 3 bugs resulting in bad buildout.cfgs.
+  [do3cc]
 
 
 0.1 (2015-09-10)